refactor(terraform): split sections to own files
