+++ enums.ts
@@ -51,43 +51,6 @@ export enum BUCKET_TYPE {
   FILE_STORAGE = 'file-storage',
 }
 
-export const getOutput = ({
-  bucketType,
-  hasStaging,
-}: {
-  bucketType: BUCKET_TYPE;
-  hasStaging: boolean;
-}): { value: string; label: string }[] => {
-  const outputs = [];
-  const bucketNames = hasStaging
-    ? [`${bucketType}-production`, `${bucketType}-staging`]
-    : [bucketType];
-  bucketNames.forEach((name) => {
-    const stageName = hasStaging && name.split('-')[name.split('-').length - 1];
-    const prefix = stageName ? `${stageName}.` : '';
-    outputs.push({
-      value: `module.${name}.AWS_SECRET_ACCESS_KEY`,
-      label: `${prefix}AWS_SECRET_ACCESS_KEY`,
-    });
-    outputs.push({
-      value: `module.${name}.AWS_ACCESS_KEY_ID`,
-      label: `${prefix}AWS_ACCESS_KEY_ID`,
-    });
-    outputs.push({
-      value: `module.${name}.BUCKET_NAME`,
-      label: `${prefix}BUCKET_NAME`,
-    });
-
-    if (bucketType === BUCKET_TYPE.WEBAPP) {
-      outputs.push({
-        value: `module.${name}.CLOUDFRONT_URL`,
-        label: `${prefix}CLOUDFRONT_URL`,
-      });
-    }
-  });
-  return outputs;
-};
-
 export enum AWS_REGIONS {
   US_EAST_2 = 'us-east-2',
   US_EAST_1 = 'us-east-1',
