+++ utils/terraform/redirect.ts
@@ -1,0 +1,22 @@
+import { TerraformProps, ModuleSpec } from './types';
+
+export const getRedirectBucketTfContent = ({
+  forwardingBucket,
+  bucketName,
+  region,
+}: TerraformProps): ModuleSpec[] => {
+  if (!forwardingBucket) {
+    return [];
+  }
+  return [
+    {
+      name: 'redirect',
+      source: `"git::https://github.com/tomfa/terraform.git//redirect"`,
+      parameters: {
+        bucket_name: `"${forwardingBucket}"`,
+        redirect_url: `"${bucketName}"`,
+        aws_region: `"${region}"`,
+      },
+    },
+  ];
+};
